Create appReducer, GlobalState, extract app display logic to UserDashboard component
